# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RUKİYE EKİNCİ Oyun Platformu API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /games:
    get:
      operationId: getGames
      tags: [games]
      summary: Get all approved games
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: orientation
          in: query
          schema:
            type: string
            enum: [vertical, horizontal]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [popular, newest, trending]
      responses:
        "200":
          description: List of games
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameListResponse"

    post:
      operationId: createGame
      tags: [games]
      summary: Upload a new game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGameRequest"
      responses:
        "201":
          description: Game created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /games/horizontal:
    get:
      operationId: getHorizontalGames
      tags: [games]
      summary: Get horizontal games for dedicated section
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of horizontal games
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameListResponse"
...
  /games/{id}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get a single game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Game details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameDetail"
        "404":
          description: Game not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteGame
      tags: [games]
      summary: Delete a game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /games/{id}/play:
    post:
      operationId: recordGamePlay
      tags: [games]
      summary: Record a game play/view
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /games/{id}/like:
    post:
      operationId: likeGame
      tags: [games]
      summary: Like or unlike a game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
  /games/{id}/comments:
    get:
      operationId: getGameComments
      tags: [games]
      summary: Get comments for a game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentListResponse"

    post:
      operationId: addGameComment
      tags: [games]
      summary: Add a comment to a game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Game:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        thumbnail:
          type: string
          nullable: true
        gameUrl:
          type: string
          nullable: true
        gameFile:
          type: string
          nullable: true
        videoPreview:
          type: string
          nullable: true
        category:
          type: string
        orientation:
          type: string
          enum: [vertical, horizontal]
        status:
          type: string
          enum: [pending, approved, rejected]
        uploaderId:
          type: integer
        uploaderName:
          type: string
        likesCount:
          type: integer
        commentsCount:
          type: integer
        playsCount:
          type: integer
        isLiked:
          type: boolean
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required: [id, title, description, category, orientation, status, uploaderId, uploaderName, likesCount, commentsCount, playsCount, isLiked, tags, createdAt]
    GameDetail:
      allOf:
        - $ref: "#/components/schemas/Game"
        - type: object
          properties:
            comments:
              type: array
              items:
                $ref: "#/components/schemas/Comment"
            relatedGames:
              type: array
              items:
                $ref: "#/components/schemas/Game"
          required: [comments, relatedGames]
    CreateGameRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        thumbnail:
          type: string
        gameUrl:
          type: string
        gameFile:
          type: string
        videoPreview:
          type: string
        category:
          type: string
        orientation:
          type: string
          enum: [vertical, horizontal]
        tags:
          type: array
          items:
            type: string
      required: [title, description, category, orientation]
    GameListResponse:
      type: object
      properties:
        games:
          type: array
          items:
            $ref: "#/components/schemas/Game"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [games, total, page, totalPages]
...
    Comment:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatar:
          type: string
          nullable: true
        likesCount:
          type: integer
        isLiked:
          type: boolean
        createdAt:
          type: string
      required: [id, content, authorId, authorName, likesCount, isLiked, createdAt]
    CommentListResponse:
      type: object
      properties:
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        total:
          type: integer
      required: [comments, total]
    CreateCommentRequest:
      type: object
      properties:
        content:
          type: string
      required: [content]
    LikeResponse:
      type: object
      properties:
        liked:
          type: boolean
        likesCount:
          type: integer
      required: [liked, likesCount]
...