# lib/api-zod/src/generated/types/createCommentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RUKİYE EKİNCİ Oyun Platformu API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCommentRequest {
  content: string;
}