# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RUKİYE EKİNCİ Oyun Platformu API
 * OpenAPI spec version: 0.1.0
 */
import type { GameOrientation } from "./gameOrientation";
import type { GameStatus } from "./gameStatus";

export interface Game {
  id: number;
  title: string;
  description: string;
  thumbnail?: string | null;
  gameUrl?: string | null;
  gameFile?: string | null;
  videoPreview?: string | null;
  category: string;
  orientation: GameOrientation;
  status: GameStatus;
  uploaderId: number;
  uploaderName: string;
  likesCount: number;
  commentsCount: number;
  playsCount: number;
  isLiked: boolean;
  tags: string[];
  createdAt: string;
}